pyproject.toml
[project]
name = "ha-renamer"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "websockets>=14.1",
]
